rename Types.ts to Type.ts
