+++ src/main/typescript/lib/metadata/Type.ts
